feat!(comp/exporter/self): rename component to clear up ambiguity (#6365)
