+++ converter/internal/staticconvert/testdata-v2/integrations_v2.river
@@ -593,10 +593,10 @@ prometheus.scrape "integrations_statsd" {
 	job_name   = "integrations/statsd"
 }
 
-prometheus.exporter.agent "integrations_agent" { }
+prometheus.exporter.self "integrations_agent" { }
 
 discovery.relabel "integrations_agent" {
-	targets = prometheus.exporter.agent.integrations_agent.targets
+	targets = prometheus.exporter.self.integrations_agent.targets
 
 	rule {
 		source_labels = ["__address__"]
