+++ converter/internal/staticconvert/testdata/sanitize.river
@@ -100,10 +100,10 @@ loki.write "logs_integrations" {
 	external_labels = {}
 }
 
-prometheus.exporter.agent "integrations_agent" { }
+prometheus.exporter.self "integrations_agent" { }
 
 discovery.relabel "integrations_agent" {
-	targets = prometheus.exporter.agent.integrations_agent.targets
+	targets = prometheus.exporter.self.integrations_agent.targets
 
 	rule {
 		source_labels = ["agent_hostname"]
